Added `convert_draw_state`
